Fix changes to KinMS API
KinMS changed its api slightly (enough to cause a major bug). In the past you had to set vPhaseCen and phaseCen in order to move the galaxy on the sky. Now vPhaseCen is set relative to phaseCen, so the old code would result in an offset. See also https://github.com/TimothyADavis/KinMSpy/issues/5